requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --index-url=https://pypi.org/simple requirements.in
#
altair==5.5.0
    # via streamlit
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
biopython==1.85
    # via -r requirements.in
blinker==1.9.0
    # via streamlit
bokeh==3.8.0
    # via -r requirements.in
cachetools==6.2.1
    # via streamlit
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via streamlit
contourpy==1.3.3
    # via bokeh
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via
    #   altair
    #   bokeh
    #   pydeck
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.9.0
    # via
    #   altair
    #   bokeh
numpy==2.3.4
    # via
    #   biopython
    #   bokeh
    #   contourpy
    #   pandas
    #   pydeck
    #   streamlit
packaging==25.0
    # via
    #   altair
    #   bokeh
    #   pytest
    #   streamlit
pandas==2.3.3
    # via
    #   -r requirements.in
    #   bokeh
    #   streamlit
pillow==11.3.0
    # via
    #   bokeh
    #   streamlit
pluggy==1.6.0
    # via pytest
protobuf==6.33.0
    # via streamlit
pyarrow==21.0.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via pytest
pytest==8.4.2
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   bokeh
    #   responses
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r requirements.in
    #   responses
    #   streamlit
responses==0.25.8
    # via -r requirements.in
rpds-py==0.28.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.50.0
    # via -r requirements.in
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via
    #   bokeh
    #   streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   referencing
    #   streamlit
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   requests
    #   responses
xyzservices==2025.4.0
    # via bokeh
